Replace DioMixin with DownloadAdapter
