+++ dio/lib/browser.dart
@@ -1,2 +1,3 @@
 export 'src/adapters/browser_adapter.dart' show BrowserHttpClientAdapter;
-export 'src/dio/dio_for_browser.dart' show DioForBrowser;
+export 'src/adapters/download/browser_download_adapter.dart' show BrowserDownloadAdapter;
+export 'src/dio.dart' show Dio;
