+++ dio/lib/io.dart
@@ -1,2 +1,3 @@
 export 'src/adapters/io_adapter.dart' hide createAdapter;
-export 'src/dio/dio_for_native.dart' show DioForNative;
+export 'src/adapters/download/io_download_adapter.dart' hide createAdapter;
+export 'src/dio.dart' show Dio;
