+++ dio/lib/src/adapters/browser_adapter.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 
 import 'package:meta/meta.dart';
 
-import '../adapter.dart';
+import 'adapter.dart';
 import '../dio_exception.dart';
 import '../headers.dart';
 import '../options.dart';
