--- dio/lib/src/dio/dio_for_browser.dart
@@ -1,37 +1,0 @@
-import '../adapters/browser_adapter.dart';
-import '../cancel_token.dart';
-import '../dio.dart';
-import '../dio_mixin.dart';
-import '../headers.dart';
-import '../options.dart';
-import '../response.dart';
-
-/// Create the [Dio] instance for Web platforms.
-Dio createDio([BaseOptions? options]) => DioForBrowser(options);
-
-/// Implements features for [Dio] on Web platforms.
-class DioForBrowser extends DioMixin implements Dio {
-  /// Create Dio instance with default [Options].
-  /// It's mostly just one Dio instance in your application.
-  DioForBrowser([BaseOptions? options]) {
-    this.options = options ?? BaseOptions();
-    httpClientAdapter = BrowserHttpClientAdapter();
-  }
-
-  @override
-  Future<Response> download(
-    String urlPath,
-    dynamic savePath, {
-    ProgressCallback? onReceiveProgress,
-    Map<String, dynamic>? queryParameters,
-    CancelToken? cancelToken,
-    bool deleteOnError = true,
-    String lengthHeader = Headers.contentLengthHeader,
-    Object? data,
-    Options? options,
-  }) {
-    throw UnsupportedError(
-      'The download method is not available in the Web environment.',
-    );
-  }
-}
