+++ example/lib/extend_dio.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:dio/io.dart';
 
-class HttpService extends DioForNative {
+class HttpService extends Dio {
   HttpService([BaseOptions? baseOptions]) : super(baseOptions) {
     options
       ..baseUrl = 'https://httpbin.org/'
